Moved person components to a new folder
